Updated packages and files
So, the issue was the model and the outdated packages. Juypter Notebook was out dated and the modules in it too. But the virtual env used to build the UI was not, so there was some naming conflicts with the outdated and updated packages, so thing had to be updated, re-run, re-built and re-deployed.